refactor : :rainbow: label-create box 위치 변경 및 필요없는 파일 제거
